HookをReact Contextを使う形に修正
