+++ example-app/SantokuApp/src/components/contexts/theme/AppThemeContext.test.tsx
@@ -2,7 +2,7 @@ import {render} from '@testing-library/react-native';
 import React from 'react';
 import {Text} from 'react-native';
 
-import {AppThemeProvider} from './AppThemeProvider';
+import {AppThemeProvider} from './AppThemeContext';
 
 describe('AppThemeProvider', () => {
   const ChildComponent: React.FC = () => {
